requirements.txt
click==8.1.3
matplotlib==3.5.2
numpy==1.22.4
python-dotenv==0.21.0
setuptools==45.2.0
torch==1.11.0
tqdm==4.64.1
wget==3.2
